Move private functions
