+++ webserver/views/api.py
@@ -125,6 +125,11 @@ def get_listens(user_id):
     }})
 
 
+def insert_json(jsonlist, user):
+    payload = _convert_to_native_format(jsonlist)
+    _send_listens_to_kafka("import", _payload_to_augmented_list(payload, user))
+
+
 def _parse_int_arg(name, default=None):
     value = request.args.get(name)
     if value:
@@ -294,6 +299,34 @@ def _validate_listen(listen):
         for ambid in ambids:
             if not is_valid_uuid(ambid):
                 _log_raise_400("Artist MBID format invalid.", listen)
+
+
+def _convert_to_native_format(data):
+    """
+    Converts the imported listen-payload from the lastfm backup file
+    to the native payload format.
+    """
+    payload = []
+    for native_lis in data:
+        listen = {}
+        listen['track_metadata'] = {}
+        listen['track_metadata']['additional_info'] = {}
+
+        if 'timestamp' in native_lis and 'unixtimestamp' in native_lis['timestamp']:
+            listen['listened_at'] = native_lis['timestamp']['unixtimestamp']
+
+        if 'track' in native_lis:
+            if 'name' in native_lis['track']:
+                listen['track_metadata']['track_name'] = native_lis['track']['name']
+            if 'mbid' in native_lis['track']:
+                listen['track_metadata']['additional_info']['recording_mbid'] = native_lis['track']['mbid']
+            if 'artist' in native_lis['track']:
+                if 'name' in native_lis['track']['artist']:
+                    listen['track_metadata']['artist_name'] = native_lis['track']['artist']['name']
+                if 'mbid' in native_lis['track']['artist']:
+                    listen['track_metadata']['additional_info']['artist_mbids'] = [native_lis['track']['artist']['mbid']]
+        payload.append(listen)
+    return payload
 
 
 def _payload_to_augmented_list(payload, user_id):
